requirements.txt
# RoleForge runtime (MVP)
psycopg2-binary>=2.9.9
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
